#33 Make Ruby shebang more flexible
Previously the Ruby executable was hardcoded to a directory path. If
ruby was installed in another directory (f. i. with a Ruby version
manager) the Ruby executable does not reside in that path (with errors
coming all along).

This commit uses `env` which automagically know where to find a possible
Rube installation.

Co-authored-by: Simon Klein <simon.klein@cloudogu.com>